+++ R/claudia.R
@@ -169,10 +169,10 @@ process.dataset <- function(d, suffix) {
   if(!(d$type %in% c("quant","cc")))
       stop("dataset ",suffix,": ","type must be quant or cc")
   
-  if(d$type=="cc") {
-      if(! "s" %in% nd)
-          stop("dataset ",suffix,": ","please give s, proportion of samples who are cases")
-      if("pvalues" %in% nd && !( "MAF" %in% nd))
+  if(d$type=="cc" & "pvalues" %in% nd) {
+      if(!( "s" %in% nd))
+          stop("dataset ",suffix,": ","please give s, proportion of samples who are cases, if using p values")
+      if(!("MAF" %in% nd))
           stop("dataset ",suffix,": ","please give MAF if using p values")
       if(d$s<=0 || d$s>=1)
           stop("dataset ",suffix,": ","s must be between 0 and 1")
@@ -196,6 +196,8 @@ process.dataset <- function(d, suffix) {
    df <- approx.bf.estimates(z=d$beta/sqrt(d$varbeta),
                               V=d$varbeta, type=d$type, suffix=suffix, sdY=d$sdY)
     df$snp <- as.character(d$snp)
+    if("position" %in% nd)
+        df <- cbind(df,position=d$position)
     return(df)
   }
 
@@ -206,11 +208,16 @@ process.dataset <- function(d, suffix) {
       d$snp <- sprintf("SNP.%s",1:length(d$pvalues))
     df <- data.frame(pvalues = d$pvalues,
                      MAF = d$MAF,
+                     N=d$N,
                      snp=as.character(d$snp))    
-    colnames(df)[-3] <- paste(colnames(df)[-3], suffix, sep=".")
-    df <- subset(df, df$MAF>0 & df$pvalues>0) # all p values and MAF > 0
-    abf <- approx.bf.p(p=df$pvalues, f=df$MAF, type=d$type, N=d$N, s=d$s, suffix=suffix)
+    snp.index <- which(colnames(df)=="snp")
+    colnames(df)[-snp.index] <- paste(colnames(df)[-snp.index], suffix, sep=".")
+    keep <- which(df$MAF>0 & df$pvalues > 0) # all p values and MAF > 0
+    df <- df[keep,]
+    abf <- approx.bf.p(p=df$pvalues, f=df$MAF, type=d$type, N=df$N, s=d$s, suffix=suffix)
     df <- cbind(df, abf)
+    if("position" %in% nd)
+        df <- cbind(df,position=d$position[keep])
     return(df)  
   }
 
@@ -229,40 +236,8 @@ process.dataset <- function(d, suffix) {
 ##' coefficients should be used if available.
 ##' 
 ##' @title Bayesian finemapping analysis
-##' @param dataset a list with the following elements
-##' \describe{
-##' 
-##'   \item{pvalues}{P-values for each SNP in dataset 1}
-##'
-##'   \item{N}{Number of samples in dataset 1}
-##'
-##'   \item{MAF}{minor allele frequency of the variants}
-##'
-##' \item{beta}{regression coefficient for each SNP from dataset 1}
-##' 
-##' \item{varbeta}{variance of beta}
-##' 
-##' \item{type}{the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control}
-##'
-##' \item{s}{for a case control dataset, the proportion of samples in dataset 1 that are cases}
-##'
-##'  \item{sdY}{for a quantitative trait, the population standard deviation of the trait.  if not given, it can be estimated from the vectors of varbeta and MAF}
-##' 
-##' \item{snp}{a character vector of snp ids, optional. If present, it will be used to merge dataset1 and dataset2.  Otherwise, the function assumes dataset1 and dataset2 contain results for the same SNPs in the same order.}
-##'
-##' }
-##'
-##' Some of these items may be missing, but you must give
-##' \itemize{
-##' \item{always}{\code{type}}
-##' \item{if \code{type}=="cc"}{\code{s}}
-##' \item{if \code{type}=="quant" and \code{sdY} known}{\code{sdY}}
-##' \item{if \code{type}=="quant" and \code{sdY} unknown}{\code{beta}, \code{varbeta}, \code{N}, \code{MAF}}
-##' and then either
-##' \item{}{\code{pvalues}, \code{MAF}}
-##' \item{}{\code{beta}, \code{varbeta}}
-##' }
-##' 
+##' @param dataset a list with specifically named elements defining the dataset
+##'   to be analysed. See \code{\link{check_dataset}} for details.
 ##'
 ##' @param p1 prior probability a SNP is associated with the trait 1, default 1e-4
 ##' @return a \code{data.frame}:
@@ -273,12 +248,11 @@ process.dataset <- function(d, suffix) {
 ##' @export
 finemap.abf <- function(dataset, p1=1e-4) {
 
-  if(!is.list(dataset))
-    stop("dataset must be a list.")
+    check_dataset(dataset,"")
   
     df <- process.dataset(d=dataset, suffix="")
     nsnps <- nrow(df)
-  dfnull <- df[1,]
+    dfnull <- df[1,]
     for(nm in colnames(df))
         dfnull[,nm] <- NA
     dfnull[,"snp"] <- "null"
@@ -315,41 +289,8 @@ finemap.abf <- function(dataset, p1=1e-4) {
 ##' coefficients should be used if available.
 ##' 
 ##' @title Fully Bayesian colocalisation analysis using Bayes Factors
-##' @param dataset1 a list with the following elements
-##' \describe{
-##' 
-##'   \item{pvalues}{P-values for each SNP in dataset 1}
-##'
-##'   \item{N}{Number of samples in dataset 1}
-##'
-##'   \item{MAF}{minor allele frequency of the variants}
-##'
-##' \item{beta}{regression coefficient for each SNP from dataset 1}
-##' 
-##' \item{varbeta}{variance of beta}
-##' 
-##' \item{type}{the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control}
-##'
-##' \item{s}{for a case control dataset, the proportion of samples in dataset 1 that are cases}
-##'
-##'  \item{sdY}{for a quantitative trait, the population standard deviation of the trait.  if not given, it can be estimated from the vectors of varbeta and MAF}
-##' 
-##' \item{snp}{a character vector of snp ids, optional. If present, it will be used to merge dataset1 and dataset2.  Otherwise, the function assumes dataset1 and dataset2 contain results for the same SNPs in the same order.}
-##'
-##' }
-##'
-##' Some of these items may be missing, but you must give
-##' \itemize{
-##' \item{always}{\code{type}}
-##' \item{if \code{type}=="cc"}{\code{s}}
-##' \item{if \code{type}=="quant" and \code{sdY} known}{\code{sdY}}
-##' \item{if \code{type}=="quant" and \code{sdY} unknown}{\code{beta}, \code{varbeta}, \code{N}, \code{MAF}}
-##' and then either
-##' \item{}{\code{pvalues}, \code{MAF}}
-##' \item{}{\code{beta}, \code{varbeta}}
-##' }
-##' 
-##'
+##' @param dataset1 a list with specifically named elements defining the dataset
+##'   to be analysed. See \code{\link{check_dataset}} for details.
 ##' @param dataset2 as above, for dataset 2
 ##' @param MAF Common minor allele frequency vector to be used for both dataset1 and dataset2, a shorthand for supplying the same vector as parts of both datasets
 ##' @param p1 prior probability a SNP is associated with trait 1, default 1e-4
@@ -358,25 +299,25 @@ finemap.abf <- function(dataset, p1=1e-4) {
 ##' @return a list of two \code{data.frame}s:
 ##' \itemize{
 ##' \item summary is a vector giving the number of SNPs analysed, and the posterior probabilities of H0 (no causal variant), H1 (causal variant for trait 1 only), H2 (causal variant for trait 2 only), H3 (two distinct causal variants) and H4 (one common causal variant)
-##' \item results is an annotated version of the input data containing log Approximate Bayes Factors and intermediate calculations, and the posterior probability SNP.PP.H4 of the SNP being causal for the shared signal
+##' \item results is an annotated version of the input data containing log Approximate Bayes Factors and intermediate calculations, and the posterior probability SNP.PP.H4 of the SNP being causal for the shared signal *if* H4 is true. This is only relevant if the posterior support for H4 in summary is convincing.
 ##' }
 ##' @author Claudia Giambartolomei, Chris Wallace
 ##' @export
 coloc.abf <- function(dataset1, dataset2, MAF=NULL, 
                       p1=1e-4, p2=1e-4, p12=1e-5) {
 
-  if(!is.list(dataset1) || !is.list(dataset2))
-    stop("dataset1 and dataset2 must be lists.")
-  if(!("MAF" %in% names(dataset1)) & !is.null(MAF))
-    dataset1$MAF <- MAF
-  if(!("MAF" %in% names(dataset2)) & !is.null(MAF))
-    dataset2$MAF <- MAF
-  
+    if(!("MAF" %in% names(dataset1)) & !is.null(MAF))
+        dataset1$MAF <- MAF
+    if(!("MAF" %in% names(dataset2)) & !is.null(MAF))
+        dataset2$MAF <- MAF
+    check_dataset(d=dataset1,1)
+    check_dataset(d=dataset2,2)
+    
   df1 <- process.dataset(d=dataset1, suffix="df1")
   df2 <- process.dataset(d=dataset2, suffix="df2")
   merged.df <- merge(df1,df2)
 
-   if(!nrow(merged.df))
+  if(!nrow(merged.df))
     stop("dataset1 and dataset2 should contain the same snps in the same order, or should contain snp names through which the common snps can be identified")
 
   merged.df$internal.sum.lABF <- with(merged.df, lABF.df1 + lABF.df2)
@@ -391,78 +332,10 @@ coloc.abf <- function(dataset1, dataset2, MAF=NULL,
   common.snps <- nrow(merged.df)
   results <- c(nsnps=common.snps, pp.abf)
   
-  output<-list(summary=results, results=merged.df)
-  return(output)
+    output<-list(summary=results,
+                 results=merged.df,
+                 priors=c(p1=p1,p2=p2,p12=p12))
+    class(output) <- c("coloc_abf",class(output))
+    return(output)
 }
 
-##' Bayesian colocalisation analysis using data.frames
-##'
-##' Converts genetic data to snpStats objects, generates p values via score tests, then runs \code{\link{coloc.abf}}
-##' 
-##' @title Bayesian colocalisation analysis using data.frames
-##' @param df1 dataset 1
-##' @param df2 dataset 2
-##' @param snps col.names for snps
-##' @param response1 col.name for response in dataset 1
-##' @param response2 col.name for response in dataset 2
-##' 
-##' @param ... parameters passed to \code{\link{coloc.abf.snpStats}}
-##' @return output of \code{\link{coloc.abf}}
-##' @export
-##' @author Chris Wallace
-coloc.abf.datasets <- function(df1,df2,
-                               snps=intersect(setdiff(colnames(df1),response1),
-                                 setdiff(colnames(df2),response2)),
-                               response1="Y", response2="Y", ...) {
-  if(length(snps)<2)
-    stop("require at least two SNPs in common between df1 and df2 to do anything sensible")
-  if(!(response1 %in% colnames(df1)))
-    stop(paste("response1",response1,"not found in df1"))
-  if(!(response2 %in% colnames(df2)))
-    stop(paste("response2",response2,"not found in df2"))
-  X1 <- new("SnpMatrix",as.matrix(df1[,snps]))
-  X2 <- new("SnpMatrix",as.matrix(df2[,snps]))
-  coloc.abf.snpStats(X1,X2,df1[,response1], df2[,response2], ...)
-}
-##' Bayesian colocalisation analysis using snpStats objects
-##'
-##' Generates p values via score tests, then runs \code{\link{coloc.abf}}
-##' @title Bayesian colocalisation analysis using snpStats objects
-##' @param X1 genetic data for dataset 1
-##' @param X2 genetic data for dataset 2
-##' @param Y1 response for dataset 1
-##' @param Y2 response for dataset 2
-##' @param snps optional subset of snps to use
-##' @param type1 type of data in Y1, "quant" or "cc"
-##' @param type2 type of data in Y2, "quant" or "cc"
-##' @param s1 the proportion of samples in dataset 1 that are cases (only relevant for case control samples)
-##' @param s2 the proportion of samples in dataset 2 that are cases (only relevant for case control samples)
-##' @param ... parameters passed to \code{\link{coloc.abf}}
-##' @return output of \code{\link{coloc.abf}}
-##' @export
-##' @author Chris Wallace
-coloc.abf.snpStats <- function(X1,X2,Y1,Y2,snps=intersect(colnames(X1),colnames(X2)),
-                               type1=c("quant","cc"),type2=c("quant","cc"),s1=NA,s2=NA,...) {
-  type1 <- match.arg(type1)
-  type2 <- match.arg(type2)
-  if(!is(X1,"SnpMatrix") || !is(X2,"SnpMatrix"))
-    stop("X1 and X2 must be SnpMatrices")
-  if(length(Y1) != nrow(X1) || length(Y2) != nrow(X2))
-    stop("length(Y1) != nrow(X1) || length(Y2) != nrow(X2)")
-  if(length(snps)<2)
-    stop("require at least two SNPs in common between X1 and X2 to do anything sensible")
-  X1 <- X1[,snps]
-  X2 <- X2[,snps]
-  if(is.null(rownames(X1)))
-    rownames(X1) <- paste("X1",1:nrow(X1),sep=".")
-  if(is.null(rownames(X2)))
-    rownames(X2) <- paste("X2",1:nrow(X2),sep=".")
-  
-  pval1 <- snpStats::p.value(single.snp.tests(phenotype=Y1, snp.data=X1),df=1)
-  pval2 <- snpStats::p.value(single.snp.tests(phenotype=Y2, snp.data=X2),df=1)
-  maf1 <- col.summary(X1)[,"MAF"]
-  maf2 <- col.summary(X2)[,"MAF"]
-  
-  coloc.abf(dataset1=list(pvalues=pval1, N=nrow(X1), MAF=maf1, snp=snps, type=type1, s=s1),
-            dataset2=list(pvalues=pval2, N=nrow(X2), MAF=maf2, snp=snps, type=type2, s=s2))
-}