+++ tests/testthat/test-abf.R
@@ -12,7 +12,7 @@ eff<-snp.rhs.estimates(Y ~ 1, snp.data=X, family="gaussian")
 beta.q <- sapply(eff@.Data, "[[", "beta")
 vbeta.q <- sapply(eff@.Data, "[[", "Var.beta")
 p.q <- pchisq(beta.q^2/vbeta.q,df=1,lower.tail=FALSE)
-sd.est <- coloc:::sdY.est(vbeta=vbeta.q, maf=maf, n=nrow(X))
+sd.est <- suppressWarnings(coloc:::sdY.est(vbeta=vbeta.q, maf=maf, n=nrow(X)))
 
 ## case-control trait
 cc <- rbinom(nrow(X),1, p=(1+as(X[,8],"numeric"))/4)
@@ -23,7 +23,7 @@ p.cc <- pchisq(beta.cc^2/vbeta.cc,df=1,lower.tail=FALSE)
 
 ## general things
 test_that("sdY.est", {
-  expect_that(abs(sd.est - sd(Y)) < 0.1, is_true())
+  expect_true(abs(sd.est - sd(Y)) < 0.1)
 })
 DQ <- list(beta=beta.q,
            varbeta=vbeta.q,
@@ -59,10 +59,10 @@ PCC.bad <- list(pvalues=p.cc,
             N=nrow(X))
 
 
-RESULTS <- list(dd = coloc.abf(dataset1=DQ,dataset2=DCC),
-                dp = coloc.abf(dataset1=DQ,dataset2=PCC),
-                pd = coloc.abf(dataset1=PQ,dataset2=DCC),
-                pp = coloc.abf(dataset1=PQ,dataset2=PCC))
+RESULTS <- suppressWarnings(list(dd = coloc.abf(dataset1=DQ,dataset2=DCC),
+                                 dp = coloc.abf(dataset1=DQ,dataset2=PCC),
+                                 pd = coloc.abf(dataset1=PQ,dataset2=DCC),
+                                 pp = coloc.abf(dataset1=PQ,dataset2=PCC)))
 lapply(RESULTS,"[[","summary")
 
 test_that("process.dataset", {
@@ -83,8 +83,8 @@ test_that("process.dataset", {
 
 ## coloc.abf with coefficients
 test_that("coloc.abf", {
-    expect_error(coloc.abf(dataset1=DQ,dataset2=DCC), NA)
-    result <- coloc.abf(dataset1=DQ,dataset2=DCC)
+    expect_error(suppressWarnings(coloc.abf(dataset1=DQ,dataset2=DCC), NA))
+    expect_warning(result <- coloc.abf(dataset1=DQ,dataset2=DCC), "minimum p value")
     expect_true(which.max(result$summary[-1]) == 5)
     expect_true(result$summary[1] == ncol(X))
 })